+++ src/renderer/index.js
@@ -8,18 +8,19 @@ require('./keyboard/bindings')
 
 const Cucumber = require('@cucumber/cucumber')
 
-const Output = require('./output')
+const MultiWritable = require('./output')
 
 const STATUS_SUCCESS = 0
 const STATUS_ERROR_DURING_CUCUMBER_RUN = 2
-const STATUS_UNCAUGHT_ERROR = 6
+const STATUS_UNCAUGHT_ERROR = 3
 
-const output = new Output({ isTTY: options.isTTY })
+const stdout = new MultiWritable({ isTTY: options.isTTY, streamName: 'stdout' })
+const stderr = new MultiWritable({ isTTY: options.isTTY, streamName: 'stderr' })
 
 const { ipcRenderer: ipc } = electron
 
 const exitWithUncaughtError = reason => {
-  output.write(reason.stack)
+  stderr.write(reason.stack)
   exitWithCode(STATUS_UNCAUGHT_ERROR)
 }
 
@@ -40,7 +41,9 @@ ipc.on('run-cucumber', () => {
     const cli = new Cucumber.Cli({
       argv: options.cucumberArgv,
       cwd: process.cwd(),
-      stdout: output,
+      stdout: stdout,
+      // TODO: Cucumber.js 6.2.1 ignores the stderr option. We should fix this.
+      stderr: stderr,
     })
     // sadly, we have to exit immediately, we can't wait for the event loop
     // to drain https://github.com/electron/electron/issues/2358